WIP #44 ADD data to registry, city selection scene, table scene
MOVE table to own scene
