+++ src/scenes/data-registry/getLogic.ts
@@ -1,0 +1,7 @@
+import { Data } from "phaser";
+import { ILogic } from "../i-logic";
+import { KEYS } from "../keys";
+
+export const getLogic = (registryHolder: {
+    registry: Data.DataManager;
+}): ILogic => registryHolder.registry.get(KEYS.registry.logic);
