+++ src/scenes/mainScene.ts
@@ -1,27 +1,21 @@
 import { Scene } from "phaser";
-import { Color } from "../Color";
 import { gameConfig } from "../game-config";
 import { logicConfig } from "../logic.config";
+import { CitySelectionScene } from "./CitySelectionScene";
+import { getLogic } from "./data-registry/getLogic";
 import { ILogic } from "./i-logic";
 import { ICity } from "./ICity";
-import { IPlayer } from "./IPlayer";
+import { KEYS } from "./keys";
 import { LogicBuilder } from "./logicBuilder";
-import { TextBuyPrice } from "./TextBuyPrice";
-import { TextPlayerMoney } from "./TextPlayerMoney";
-import { TextSellPrice } from "./TextSellPrice";
-import { WareType } from "./wareType";
+import { TableScene } from "./TableScene";
 
 export class MainScene extends Scene {
     private logic!: ILogic;
     private city!: ICity;
-    private player!: IPlayer;
-    private playerMoneyText!: TextPlayerMoney;
-    private textBuyPrices: TextBuyPrice[] = [];
-    private textSellPrices: TextSellPrice[] = [];
 
     constructor() {
         super({
-            key: "MainScene",
+            key: KEYS.scenes.main,
         });
     }
 
@@ -37,15 +31,20 @@ export class MainScene extends Scene {
 
     public create(): void {
         const logicObjects = LogicBuilder.create();
-        this.logic = logicObjects.logic;
-        this.city = logicObjects.city;
-        this.player = logicObjects.player;
+        this.registry.set(KEYS.registry.logic, logicObjects.logic);
+        this.registry.set(KEYS.registry.cities, logicObjects.cities);
+        this.registry.set(KEYS.registry.player, logicObjects.player);
+        this.logic = getLogic(this);
+        this.city = this.logic.city;
+
         this.addBackgroundMusic();
         this.addBackground();
-        this.addPlayerMoneyText(logicObjects.player);
-        this.addTable();
+
+        this.scene.add(KEYS.scenes.citySelection, CitySelectionScene, true);
+        this.scene.add(KEYS.scenes.table, TableScene, true);
 
         this.time.addEvent({
+            // TODO #44 adapt to multi city
             callback: () => this.city.consume(),
             delay: logicConfig.cityConsumeTime,
             loop: true,
@@ -53,10 +52,6 @@ export class MainScene extends Scene {
     }
 
     public update() {
-        const update = (x: { update(): void }) => x.update();
-        this.playerMoneyText.update();
-        this.textBuyPrices.forEach(update);
-        this.textSellPrices.forEach(update);
         if (this.logic.gameOver()) {
             this.sound.removeByKey("background");
             this.scene.start("EndScene");
@@ -76,118 +71,4 @@ export class MainScene extends Scene {
                 (gameConfig.height as number) / 300
             );
     }
-
-    private addPlayerMoneyText(player: IPlayer) {
-        this.playerMoneyText = new TextPlayerMoney(this, 50, 75, "", {});
-        this.children.add(this.playerMoneyText);
-        this.playerMoneyText.init(player);
-    }
-
-    private addTable() {
-        this.addHeader();
-        Object.values(WareType).forEach((ware, i) => {
-            this.addRow(ware, 200 + i * 150);
-        });
-    }
-    private addHeader() {
-        const addTextAtY = this.addTableText(150);
-        addTextAtY(200, "City");
-        addTextAtY(300, "You");
-        addTextAtY(600, "Buy");
-        addTextAtY(700, "Sell");
-    }
-
-    private addRow(ware: WareType, y: number) {
-        this.addBuyButton(ware, y);
-        this.addSellButton(ware, y);
-        const addTextAtY = this.addTableText(y);
-        addTextAtY(50, ware);
-        this.addCityQuantityText(addTextAtY, ware);
-        this.addPlayerQuantityText(addTextAtY, ware);
-        this.addBuyPrice(y, ware);
-        this.addSellPrice(y, ware);
-    }
-
-    private addBuyPrice(y: number, ware: WareType) {
-        const text = new TextBuyPrice(this, 600, y, "", {});
-        this.textBuyPrices.push(text);
-        this.children.add(text);
-        text.init(this.city, ware);
-    }
-    private addSellPrice(y: number, ware: WareType) {
-        const text = new TextSellPrice(this, 700, y, "", {});
-        this.textSellPrices.push(text);
-        this.children.add(text);
-        text.init(this.city, ware);
-    }
-    /*interessant*/
-    private addBuyButton(ware: WareType, y: number) {
-        this.addButton("buy", { x: 400, y }, () => this.logic.buy(ware));
-    }
-
-    private addSellButton(ware: WareType, y: number) {
-        this.addButton("sell", { x: 500, y }, () => this.logic.sell(ware));
-    }
-
-    private addCityQuantityText(
-        addTextAtY: (x: number, text: string) => Phaser.GameObjects.Text,
-        ware: WareType
-    ) {
-        const cityQuantityText = addTextAtY(
-            200,
-            this.city
-                .get(ware)
-                .getQuantity()
-                .toString()
-        );
-        this.city
-            .get(ware)
-            .getStream()
-            .subscribe(quantity =>
-                cityQuantityText.setText(quantity.toString())
-            );
-    }
-
-    private addPlayerQuantityText(
-        addTextAtY: (x: number, text: string) => Phaser.GameObjects.Text,
-        ware: WareType
-    ) {
-        const playerQuantityText = addTextAtY(
-            300,
-            this.player
-                .get(ware)
-                .getQuantity()
-                .toString()
-        );
-        this.player
-            .get(ware)
-            .getStream()
-            .subscribe(quantity =>
-                playerQuantityText.setText(quantity.toString())
-            );
-    }
-
-    private addTableText(y: number) {
-        return (x: number, text: string) =>
-            this.add
-                .text(x, y, text)
-                .setFontFamily("Arial")
-                .setFontSize(32)
-                .setColor(Color.black);
-    }
-
-    private addButton(key: string, pos: IPoint, logicCallback: () => void) {
-        const button = this.add.sprite(pos.x, pos.y, key);
-        button.setInteractive();
-        const callBackWithSound = () => {
-            logicCallback();
-            this.sound.play(key);
-        };
-        button.on("pointerdown", callBackWithSound);
-    }
-}
-
-interface IPoint {
-    x: number;
-    y: number;
 }
